refactor: single file exports matcher
BREAKING CHANGE:
toPassPackageAudit is no longer automatically extended onto jest, using setupFilesAfterEnv will no longer work.
