+++ source/__mocks__/mock-spawn.d.ts
